requirements.txt
email-validator==2.2.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
SendGrid==6.10.0
SQLAlchemy==2.0.29
Werkzeug==3.0.1
cairosvg
pillow
reportlab
svglib
